pyproject.toml
[project]
name = "smiles-iupac"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "batchata",
    "datasets>=4.5.0",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "polars>=1.37.1",
    "py2opsin>=1.1.0",
    "python-dotenv>=1.2.1",
    "rdkit>=2025.9.4",
]

[tool.uv.sources]
batchata = { git = "https://github.com/lukasmki/batchata.git" }
